Move LnUrlPayError to sdk-common
