+++ libs/sdk-core/src/binding.rs
@@ -24,8 +24,8 @@ use tokio::sync::Mutex;
 use crate::breez_services::{self, BreezEvent, BreezServices, EventListener};
 use crate::chain::RecommendedFees;
 use crate::error::{
-    ConnectError, LnUrlAuthError, LnUrlPayError, LnUrlWithdrawError, ReceiveOnchainError,
-    ReceivePaymentError, SdkError, SendOnchainError, SendPaymentError,
+    ConnectError, LnUrlAuthError, LnUrlWithdrawError, ReceiveOnchainError, ReceivePaymentError,
+    SdkError, SendOnchainError, SendPaymentError,
 };
 use crate::fiat::{FiatCurrency, Rate};
 use crate::lsp::LspInformation;
