+++ libs/sdk-core/src/breez_services.rs
@@ -32,9 +32,8 @@ use crate::chain::{
     DEFAULT_MEMPOOL_SPACE_URL,
 };
 use crate::error::{
-    ConnectError, LnUrlAuthError, LnUrlPayError, LnUrlWithdrawError, ReceiveOnchainError,
-    ReceiveOnchainResult, ReceivePaymentError, SdkError, SdkResult, SendOnchainError,
-    SendPaymentError,
+    ConnectError, LnUrlAuthError, LnUrlWithdrawError, ReceiveOnchainError, ReceiveOnchainResult,
+    ReceivePaymentError, SdkError, SdkResult, SendOnchainError, SendPaymentError,
 };
 use crate::fiat::{FiatCurrency, Rate};
 use crate::greenlight::{GLBackupTransport, Greenlight};
